+++ Shared/MovappApp.swift
@@ -11,13 +11,11 @@ import SwiftUI
 struct MovappApp: App {
 
     let dictionaryDataStore = DictionaryDataStore.shared
-    var forChildrenDataStore: ForChildrenDataStore { ForChildrenDataStore(dictionaryDataStore: dictionaryDataStore) }
+    let forChildrenDataStore: ForChildrenDataStore
     let userDefaultsStore = UserDefaultsStore()
     let teamDataStore = TeamDataStore()
 
-    var languageStore: LanguageStore { LanguageStore(userDefaultsStore: userDefaultsStore,
-                                                      dictionaryDataStore: dictionaryDataStore,
-                                                      forChildrenDataStore: forChildrenDataStore) }
+    let languageStore: LanguageStore
     let soundService = SoundService()
     var favoritesProvider: PhrasesFavoritesProvider { PhrasesFavoritesProvider(favoritesService: favoritesService) }
     let favoritesService: PhraseFavoritesService
@@ -45,9 +43,14 @@ struct MovappApp: App {
         self.onBoardingDataStore = OnBoardingStore(userDefaultsStore: userDefaultsStore)
         self.favoritesService = PhraseFavoritesService(userDefaultsStore: userDefaultsStore,
                                                        dictionaryDataStore: dictionaryDataStore)
+        self.forChildrenDataStore = ForChildrenDataStore(dictionaryDataStore: dictionaryDataStore)
+        self.languageStore = LanguageStore(userDefaultsStore: userDefaultsStore,
+                                           dictionaryDataStore: dictionaryDataStore,
+                                           forChildrenDataStore: forChildrenDataStore)
 
         // Restoring application language from locale based on Fastlane snapshots
-        if let appleLocale = UserDefaults.standard.string(forKey: "AppleLocale") {
+        if UserDefaults.standard.value(forKey: "FASTLANE_SNAPSHOT") != nil,
+            let appleLocale = UserDefaults.standard.string(forKey: "AppleLocale") {
             userDefaultsStore.storeLanguage(getSetLanguage(for: appleLocale))
             NSLog("Restored language from arguments: \(appleLocale)")
         }