+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Details about this file’s format at <http://keepachangelog.com/>. The change log is parsed automatically when minting releases through Fastlane, see `Fastlane/Fastfile`.
 
 ## [Unreleased]
+- Fix language switching
 
 ## [1.8.0, build 72] - 2023-07-23Z
 - #211 Fixed adding to favourite phrases